Plugin abstract schema.
