+++ source/abstracts/dynamics/pluginschema.cpp
@@ -1,0 +1,5 @@
+#include "moduleschema.hpp"
+
+TEGRA_USING_NAMESPACE Tegra;
+TEGRA_NAMESPACE_BEGIN(Tegra::Abstracts)
+TEGRA_NAMESPACE_END
